feat(List)(shared):  add PercentageRow component
